organised the files, every page code is now in their respective pages folder/file
