--- src/healers-healthcare-frontend/src/pages/Appointments/Appointments.jsx
@@ -1,12 +1,0 @@
-import Appointment from "../../components/AppointmentsComponent1";
-
-function Appointments() {
-  return (
-    <div>
-      {/* <h1>Appointments</h1> */}
-      <Appointment />
-    </div>
-  );
-}
-
-export default Appointments;
