--- src/healers-healthcare-frontend/src/pages/DoctorDashboard/DoctorDashboard.jsx
@@ -1,11 +1,0 @@
-import DoctorRecord from "../../components/DoctorComponent1";
-
-function DoctorDashboard() {
-  return (
-    <div>
-      <DoctorRecord />
-    </div>
-  );
-}
-
-export default DoctorDashboard;
