--- src/healers-healthcare-frontend/src/pages/EhrPage/EhrPage.jsx
@@ -1,16 +1,0 @@
-import Component1 from "../../components/EhrComponent1";
-
-
-
-function Ehr () {
-  return (
-    <div>
-      
-                
-      <Component1 />
-    </div>
-  );
-}
-
-
-export default Ehr;
