--- src/healers-healthcare-frontend/src/pages/Inventory/InventoryPage.jsx
@@ -1,12 +1,0 @@
-import Inventory1 from "../../components/InventoryComponent1";
-
-
-function InventoryPage() {
-  return (
-    <div>
-    <Inventory1 />
-    </div>
-  );
-}
-
-export default InventoryPage;
